+++ Common/PresentationLayer/Models/DepartmentModel.vb
@@ -8,7 +8,7 @@
         Public Property DepartmentNameAra() As String
         Public Property ParentIdNo() As Integer?
         Public Property Notes() As String
-        Public Property RevCostCenterIdNo() As Integer
+        Public Property RevCostCenterIdNo() As Int16
         Public Property SortKey As String
     End Class
